Clean up Flights table columns
Have Duration be of type string instead of time, so the presentation is
cleaner and the information is more direct and controlled. Rather than
respresenting as milliseconds since epoch, just say the flight lasts
"8 hours and 24 minutes."

Also, the foreign keys for the two airports are now set to null: false
like they're supposed to be, but I forgor when I made them earlier.