Expand database to prepare for flight bookings
Database has been expended with Booking and Passenger tables to get
ready for the next step after searching up flights. Searching for
flights results in radio lists of Airport objects currently, but the
idea is that they will be flights that allow the user to submit a
subform that will create booking and passenger objects.